requirements.txt
nonebot2
httpx
openai
pillow>=10.0
pyoverload